Break out Repo access from API.ReleaseController
